Make cutting section fixture use case for testing accessible for all implementations of TrackDataset
