+++ tests/OTAnalytics/plugin_datastore/test_track_store.py
@@ -8,12 +8,11 @@ from pandas import DataFrame, Series
 from OTAnalytics.domain import track
 from OTAnalytics.domain.event import Event
 from OTAnalytics.domain.geometry import (
-    Coordinate,
     ImageCoordinate,
     RelativeOffsetCoordinate,
     calculate_direction_vector,
 )
-from OTAnalytics.domain.section import LineSection, SectionId
+from OTAnalytics.domain.section import LineSection
 from OTAnalytics.domain.track import Track, TrackId
 from OTAnalytics.domain.track_dataset import (
     TRACK_GEOMETRY_FACTORY,
@@ -40,59 +39,11 @@ from tests.conftest import (
     assert_equal_detection_properties,
     assert_equal_track_properties,
     assert_track_datasets_equal,
-    create_track,
 )
 from tests.OTAnalytics.plugin_datastore.conftest import (
     assert_track_geometry_dataset_add_all_called_correctly,
     create_mock_geometry_dataset,
 )
-
-
-@pytest.fixture
-def cutting_section_test_case() -> (
-    tuple[LineSection, list[Track], list[Track], set[TrackId]]
-):
-    first_track = create_track(
-        "1",
-        [(1, 1), (2, 1), (3, 1), (4, 1), (4, 2), (3, 2), (2, 2), (1, 2)],
-        start_second=1,
-    )
-    expected_first_track_1 = create_track(
-        "1_1",
-        [
-            (1, 1),
-            (2, 1),
-        ],
-        1,
-    )
-    expected_first_track_2 = create_track("1_2", [(3, 1), (4, 1), (4, 2), (3, 2)], 3)
-    expected_first_track_3 = create_track("1_3", [(2, 2), (1, 2)], 7)
-
-    second_track = create_track("2", [(1, 1), (2, 1), (3, 1)], 1)
-    expected_second_track_1 = create_track("2_1", [(1, 1), (2, 1)], 1)
-    expected_second_track_2 = create_track("2_2", [(3, 1)], 3)
-
-    third_track = create_track("3", [(10, 10), (20, 10)], 10)
-
-    _id = "#cut_1"
-    cutting_section = LineSection(
-        SectionId(_id), _id, {}, {}, [Coordinate(2.5, 0), Coordinate(2.5, 3)]
-    )
-
-    expected_original_track_ids = {first_track.id, second_track.id}
-
-    return (
-        cutting_section,
-        [first_track, second_track, third_track],
-        [
-            expected_first_track_1,
-            expected_first_track_2,
-            expected_first_track_3,
-            expected_second_track_1,
-            expected_second_track_2,
-        ],
-        expected_original_track_ids,
-    )
 
 
 class TestPandasDetection:
